screw you *c++s your codebase*
