+++ main/include/config.hpp
@@ -1,0 +1,4 @@
+#include <stdbool.h>
+#include <string>
+
+struct config {};
