+++ main/include/connection_manager.h
@@ -1,0 +1,4 @@
+
+void init_wifi_ap(void);
+void init_wifi_sta(void);
+void init_mdns(void);
